+++ CHANGELOG.md
@@ -6,6 +6,10 @@
   session
 - Added new `SType::Nasdaq` and `SType::Cms` to support querying US equities datasets
   using either convention, regardless of the original convention of the dataset.
+- Relaxed `pyo3`, `tokio`, and `zstd` dependency version requirements
+
+### Bug fixes
+- Fixed multi-frame Zstd decoding for async decoders
 
 ## 0.8.0 - 2023-07-19
 ### Enhancements